積ん読対象コントローラにてsuperで呼び出していたコードをモジュール化
